[WIP] move event wrapper types into their own module
